Reorganize files in lib/base/components
