+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "sketchbook2",
-    "version": "0.0.14",
+    "version": "0.0.16",
     "private": true,
     "scripts": {
         "dev": "vite dev",
