+++ src/lib/components/Inputs/NumberInput.svelte
