+++ src/lib/components/Inputs/styles/input-color.scss
