+++ src/lib/components/Inputs/styles/input-range.scss
