+++ src/lib/components/MainView/ProjectViewer.svelte
