+++ src/lib/components/Panels/PanelFooter.svelte
