+++ src/lib/components/SettingsPanel/SettingsPanel.svelte
@@ -1,7 +1,7 @@
 <script>
     import SettingsContent from './SettingsContent.svelte';
-    import PanelHeader from '../PanelHeader.svelte';
-    import Panel from '../Panel.svelte';
+    import PanelHeader from '../Panels/PanelHeader.svelte';
+    import Panel from '../Panels/Panel.svelte';
     import { content } from '../../../config/content';
 </script>
