+++ src/lib/components/SettingsPanel/SettingsParamConfigs.ts
